requirements.txt
pandas
numpy
scikit-learn
ydata-profiling
langchain
faiss-cpu
pydantic
python-dotenv
PyYAML
tqdm
openai>=1.42.0